+++ src/Generator/ServiceWorkerGenerator.php
@@ -12,10 +12,8 @@
 namespace HeimrichHannot\ContaoPwaBundle\Generator;
 
 
-use Contao\Controller;
 use Contao\PageModel;
 use HeimrichHannot\ContaoPwaBundle\DataContainer\PageContainer;
-use HeimrichHannot\ContaoPwaBundle\Manifest\Manifest;
 use HeimrichHannot\ContaoPwaBundle\Model\PwaConfigurationsModel;
 use HeimrichHannot\UtilsBundle\Template\TemplateUtil;
 use Symfony\Bridge\Monolog\Logger;